+++ modules/alloy/src/main/ts/ephox/alloy/ui/slider/SliderUi.ts
@@ -1,4 +1,4 @@
-import { Fun, Optional } from '@ephox/katamari';
+import { Optional } from '@ephox/katamari';
 
 import { Keying } from '../../api/behaviour/Keying';
 import { Receiving } from '../../api/behaviour/Receiving';
@@ -86,6 +86,10 @@ const sketch: CompositeSketchFactory<SliderDetail, SliderSpec> = (detail: Slider
     choose(slider);
   };
 
+  const focusWidget = (component: AlloyComponent) => {
+    AlloyParts.getPart(component, detail, 'spectrum').map(Keying.focusIn);
+  };
+
   return {
     uid: detail.uid,
     dom: detail.dom,
@@ -96,9 +100,7 @@ const sketch: CompositeSketchFactory<SliderDetail, SliderSpec> = (detail: Slider
       [
         Keying.config({
           mode: 'special',
-          focusIn: (slider) => {
-            return AlloyParts.getPart(slider, detail, 'spectrum').map(Keying.focusIn).map(Fun.always);
-          }
+          focusIn: focusWidget
         }),
         Representing.config({
           store: {
@@ -138,8 +140,11 @@ const sketch: CompositeSketchFactory<SliderDetail, SliderSpec> = (detail: Slider
       }),
       AlloyEvents.run(NativeEvents.touchstart(), onDragStart),
       AlloyEvents.run(NativeEvents.touchend(), onDragEnd),
-      AlloyEvents.run(NativeEvents.mousedown(), onDragStart),
-      AlloyEvents.run(NativeEvents.mouseup(), onDragEnd)
+      AlloyEvents.run(NativeEvents.mousedown(), (component, event: NativeSimulatedEvent<DragEvent>) => {
+        focusWidget(component);
+        onDragStart(component, event);
+      }),
+      AlloyEvents.run(NativeEvents.mouseup(), onDragEnd),
     ]),
 
     apis: {