+++ modules/tinymce/src/plugins/lists/main/ts/listmodel/ParseLists.ts
@@ -1,8 +1,8 @@
 import { Arr, Cell, Optional } from '@ephox/katamari';
 import { Compare, SugarElement, SugarNode, Traverse } from '@ephox/sugar';
 
-import { createEntry, Entry, EntryComment, EntryNoList, isEntryList } from './Entry';
-import { isList } from './Util';
+import { createEntry, Entry, EntryFragment } from './Entry';
+import { isList, isListItem, ListType } from './Util';
 
 type Parser = (depth: number, itemSelection: Optional<ItemSelection>, selectionState: Cell<boolean>, element: SugarElement) => Entry[];
 
@@ -15,22 +15,6 @@ export interface EntrySet {
   readonly entries: Entry[];
   readonly sourceList: SugarElement<HTMLElement>;
 }
-
-const entryToEntryNoList = (entry: Entry, type: string, isInPreviousLi: boolean): EntryNoList | EntryComment => {
-  if (isEntryList(entry)) {
-    return ({
-      depth: entry.depth,
-      dirty: entry.dirty,
-      content: entry.content,
-      isSelected: entry.isSelected,
-      type,
-      attributes: entry.itemAttributes,
-      isInPreviousLi
-    });
-  } else {
-    return entry;
-  }
-};
 
 const parseSingleItem: Parser = (depth: number, itemSelection: Optional<ItemSelection>, selectionState: Cell<boolean>, item: SugarElement<HTMLElement>): Entry[] => {
   if (SugarNode.isComment(item)) {
@@ -70,14 +54,23 @@ const parseItem: Parser = (depth: number, itemSelection: Optional<ItemSelection>
   Traverse.firstChild(item).filter(isList).fold(
     () => parseSingleItem(depth, itemSelection, selectionState, item),
     (list) => {
-      const parsedSiblings = Arr.foldl(Traverse.children(item), (acc: Entry[], s, i) => {
+      const parsedSiblings = Arr.foldl(Traverse.children(item), (acc: Entry[], liChild, i) => {
         if (i === 0) {
           return acc;
         } else {
-          const parsedSibling = parseSingleItem(depth, itemSelection, selectionState, s)
-            .map((e) => entryToEntryNoList(e, s.dom.nodeName.toLowerCase(), true));
-
-          return acc.concat(parsedSibling);
+          if (isListItem(liChild)) {
+            return acc.concat(parseSingleItem(depth, itemSelection, selectionState, liChild));
+          } else {
+            const fragment: EntryFragment = {
+              isFragment: true,
+              depth,
+              content: [ liChild ],
+              isSelected: false,
+              dirty: false,
+              parentListType: SugarNode.name(list) as ListType
+            };
+            return acc.concat(fragment);
+          }
         }
       }, []);