+++ modules/snooker/CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Fixed
+- `TableOperations.opEraseRows` did not calculate the correct row index for colgroup tables. #TINY-6309
+- `TableOperations.opPasteColsAfter` & `TableOperations.opPasteRowsAfter` now place the cursor in the newly pasted row/column to be consistent with the other pasting operations. #TINY-6460
+
 ## 11.0.9 - 2023-03-15
 
 ### Fixed